data-scripts/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
beautifulsoup4==4.9.0     # via -r requirements.in
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
et-xmlfile==1.0.1         # via openpyxl
html5lib==1.0.1           # via -r requirements.in
idna==2.9                 # via requests
jdcal==1.4.1              # via openpyxl
numpy==1.18.4             # via pandas, scipy
openpyxl==3.0.3           # via -r requirements.in
pandas==1.0.3             # via -r requirements.in
python-dateutil==2.8.1    # via pandas
pytz==2020.1              # via pandas
requests==2.23.0          # via -r requirements.in
scipy==1.4.1              # via -r requirements.in
six==1.14.0               # via html5lib, python-dateutil
soupsieve==2.0            # via beautifulsoup4
urllib3==1.25.9           # via requests
webencodings==0.5.1       # via html5lib
wheel==0.34.2             # via -r requirements.in
xlrd==1.2.0               # via -r requirements.in
